requirements-realtime.txt
# Optional dependencies for real-time integration
# Install with: pip install -r requirements-realtime.txt

# WebSocket client
websockets>=10.0

# HTTP async client
aiohttp>=3.8.0

# MQTT protocol
paho-mqtt>=1.7.0

# Environment variable management
python-dotenv>=0.20.0

# Optional: for monitoring
prometheus-client>=0.14.0

# Optional: structured logging
python-json-logger>=2.0.0
